+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 #### Fixed
 - Incorrect attribute values generated when literals are substituted in CA expressions
+- AINSERT grammar tests improved
 
 ## [1.2.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.1.0...1.2.0) (2022-05-11)